requirements.txt
hutil
pandas
numpy
pyyaml
ruamel.yaml
opencv-python
dataclasses
tqdm
torch
Pillow
clip
torchvision
simple_parsing
ftfy
regex
